Created a "Pause" Menu
Instead when the player presses M, they go to the pause menu, where they can restart the game, save the game, load the game, or go straight back to the game. Please note that only the restart the game works. And if you press the Return to the game button, it also resets the scene.